lab06 B recommit
I created 2 different content views, one for the start screen, and one for the actual game. I made it so that clicking the correct answer would increase your score, while clicking a wrong answer would end the game and take you back to the start screen. Currently there are no questions that the game asks, and the buttons just say whether they are correct or wrong.